+++ R/catchCurveSim.R
@@ -8,17 +8,17 @@
 #'
 #' The \bold{\sQuote{Z cv}} controls the coefficient of variation (as a proportion, not a percentage) to use when modeling the instantaneous mortality rate.  If \bold{\sQuote{Z cv}} is set to 0 then no randomization is used.  If \bold{\sQuote{Z cv}} is greater than 0 then a random normal deviate centered on \bold{\sQuote{Z mean}} with a standard deviation of \bold{\sQuote{Z cv}}*\bold{\sQuote{Z mean}} is used to model the instantaneous mortality rate.
 #'
-#' A simulated change in instantaneous mortality rate is controlled with the \bold{\sQuote{Z delta}} slider bar and the \code{ZdeltaAge}, and \code{ZSteady} arguments.  The \bold{\sQuote{Z.delta}} slider controls a proportional adjustment in the mortality rate that begins at \code{ZdeltaAge}.  The adjustment is a constant proportion through time/ages if \code{ZSteady=TRUE}, but grows geometrically if \code{ZSteady=FALSE}.  For example, if \bold{\sQuote{Z mean}} is set to 0.40, \bold{\sQuote{Z cv}} is set to 0 (i.e., no variability), \bold{\sQuote{Z delta}} is set to 1.2, \code{ZdeltaAge=5}, and \code{ZSteady=TRUE} then the vector of mortalities used for ages 1 to 10 would be
+#' A simulated change in instantaneous mortality rate is controlled with the \bold{\sQuote{Z delta}} slider bar, \bold{\sQuote{Steady?}} checkbox, and the \code{ZdeltaAge} arguments.  The \bold{\sQuote{Z.delta}} slider controls a proportional adjustment in the mortality rate that begins at \code{ZdeltaAge}.  The adjustment is a constant proportion through time/ages if \bold{\sQuote{Steady?}} is checked, but grows geometrically if\bold{\sQuote{Steady?}} is not checked.  For example, if \bold{\sQuote{Z mean}} is set to 0.40, \bold{\sQuote{Z cv}} is set to 0 (i.e., no variability), \bold{\sQuote{Z delta}} is set to 1.2, \code{ZdeltaAge=5}, and \bold{\sQuote{Steady?}} is checked then the vector of mortalities used for ages 1 to 10 would be
 #'
 #' \code{c(0.4,0.4,0.4,0.4,0.4*1.2,0.4*1.2,0.4*1.2,0.4*1.2,0.4*1.2,0.4*1.2)}
 #'
-#' Alternatively, if all values were the same except that \code{ZSteady=FALSE}, then the vector of mortalities used for ages 1 to 10 would be 
+#' Alternatively, if all values were the same except that \bold{\sQuote{Steady?}} is not checked, then the vector of mortalities used for ages 1 to 10 would be 
 #' 
 #' \code{c(0.4,0.4,0.4,0.4,0.4*1.2,0.4*1.2^2,0.4*1.2^3,0.4*1.2^4,0.4*1.2^5,0.4*1.2^6)}
 #'
 #' The simulation code can consider two different tacks when modeling variability in the instantaneous mortality rate. The first tack considers that the mortality rate experienced throughout the existence of each year-class corresponding to ages in the cross-sectional sample older than some defined age was different than all previous year-classes.  Simulations using this tack can be used to assess the assumption of a constant mortality rate for all year-classes.  The second tack considers that the mortality rate for all ages prior to some defined age was the same for all year-classes but that a different mortality rate for all ages after some defined age for all year-classes should be used.  Simulations using this tack can be used to assess the assumption of a constant mortality rate for all ages.  Which track is used in a set of simulations is set by the \code{Zycl=} argument to the function (and not by a slider bar).  If \code{Zcyl=TRUE} then the first tack is used.
 #'
-#' Slider bars are used to control similar aspects of the initial population size (No).  \bold{\sQuote{No mean}} and \bold{\sQuote{No CV}} are the mean and coefficient of variation to be used in the simulations.  Changes in the initial population size are simulated with the \bold{\sQuote{No delta}} slider and the \code{NodeltaAge} and \code{NoSteady} arguments.  These adjustments are applied in the same manner as described for the instantaneous mortality rate.
+#' Slider bars are used to control similar aspects of the initial population size (No).  \bold{\sQuote{No mean}} and \bold{\sQuote{No CV}} are the mean and coefficient of variation to be used in the simulations.  Changes in the initial population size are simulated with the \bold{\sQuote{No delta}} slider, \bold{\sQuote{Steady?}} check box, and the \code{NodeltaAge} argument.  These adjustments are applied in the same manner as described for the instantaneous mortality rate.
 #'
 #' The vulnerabilties of each age of fish to the fishing gear can be simulated in one of two ways.  If a vector of \dQuote{1}s is sent in the \code{v} argument (this is the default) then the fish in all ages after the recruitment age (which is selected with a slider bar) are fully and equally vulnerable to the gear.  Fish in age-classes prior to the age of recruitment are modeled by a linear increase in proportional vulnerability from a value of 0.1 to 1 (for the age at recruitment).  A wider variety of changes in vulnerabilities can be modeled by sending various vectors of proportional vulnerabilities for each age in the simulation.  For example, if 10 ages are modeled, then the vector \code{c(0.1,0.3,0.6,1,1,1,0.8,0.6,0.4,0.2)} could be used to model incomplete vulnerabilities prior to age 4, complete vulnerablities from age 4 to age 6, and then incomplete vulnerabilities for age 7 and older.  If the vector of vulnerabilities sent does not contain all \dQuote{1}s then a recruitment age slider will not be available.
 #'
@@ -27,10 +27,7 @@
 #' @param max.age A single numeric integer that indicates the oldest age to use in the simulations.
 #' @param v A single vector of length \code{max.age} of the proportional vulnerabilities to the gear.  See details.
 #' @param Zycl A single logical that indicates whether changes in mortality rates are applied to and followed for each year-class or not.  See details.
-#' @param ZdeltaAge A single numeric that indicates the age at which the mortality rate should change.  See details.
-#' @param NodeltaAge A single numeric that indicates the age at which the initial population size should change.  See details.
-#' @param ZSteady A single logical that indicates whether the change in mortality rate will be steady (\code{=TRUE}; default) or not.  See details.
-#' @param NoSteady A single logical that indicates whether the change in initialpopulation size will be steady (\code{=TRUE}; default) or not.  See details.
+#' @param deltaAge A single numeric that indicates the age at which the mortality rate and population size should change.  See details.
 #' @param recruit.age A single numeric that indicates what the age at full recruitment should be.  See detials.
 #'
 #' @return None.  A slide controlled dynamic graphic with the following description is produced.  The catch curve graphic is shown for as many as three lines.  The gray line is the catch-at-age data for the default values (the initial values when the simulation was begun).  This line is used simply as a basis for examining parameter changes.  The blue line is the catch-ag-age data for current choices of \bold{\sQuote{Z.mean}}, \bold{\sQuote{Z.cv}}, \bold{\sQuote{No.mean}}, and \bold{\sQuote{No.cv}}, but not for any of the other choices (i.e., no change in the \sQuote{delta} values).
@@ -55,63 +52,60 @@
 #' 
 #'@export
 #'
-catchCurveSim <- function(max.age=15,v=rep(1,max.age),
-                          Zycl=FALSE,ZdeltaAge=round(max.age/2,0),
-                          NodeltaAge=ZdeltaAge,ZSteady=TRUE,NoSteady=ZSteady,
-                          recruit.age=3) {
-  type <- EM <- En <- NULL
+catchCurveSim <- function(max.age=15,v=rep(1,max.age),Zycl=FALSE,
+                          deltaAge=round(max.age/2,0),recruit.age=3) {
+  # Attempt to deal with global binding issue
+  Z.mean <- No.mean <- Z.delta <- No.delta <- Z.cv <- No.cv <- steady <- NULL
   if (iChk4Namespace("manipulate")) {
     manipulate::manipulate(
       {
         set.seed(sample(1:100000))
-        Z.param <- list(mean=Z.mean,cv=Z.cv,delta=Z.delta,
-                        deltaAge=ZdeltaAge,dstdy=ZSteady)
-        No.param <- list(mean=No.mean,cv=No.cv,delta=No.delta,
-                         deltaAge=NodeltaAge,dstdy=NoSteady)    
-        iCC.plot(max.age,v,Zycl,ZdeltaAge,NodeltaAge,recruit.age,Z.param,No.param)
+        Z.param <- list(mean=Z.mean,cv=Z.cv,delta=Z.delta,deltaAge=deltaAge,dstdy=steady)
+        No.param <- list(mean=No.mean,cv=No.cv,delta=No.delta,deltaAge=deltaAge,dstdy=steady)    
+        iCC.plot(max.age,v,Zycl,deltaAge,recruit.age,Z.param,No.param)
       },
-      Z.mean=manipulate::slider(0.05,1,step=0.05,initial=0.4,label="Z mean"),
-      No.mean=manipulate::slider(100,2000,step=100,initial=1000,label="No mean"),
-      Z.cv=manipulate::slider(0,0.2,step=0.05,initial=0,label="Z CV"),
-      No.cv=manipulate::slider(0,0.5,step=0.05,initial=0,label="No CV"),
+      Z.mean=manipulate::slider(0.1,0.8,step=0.05,initial=0.4,label="Z mean"),
+      No.mean=manipulate::slider(100,1000,step=100,initial=700,label="No mean"),
       Z.delta=manipulate::slider(0.3,2,step=0.1,initial=1,label="Z delta"),
       No.delta=manipulate::slider(0.3,2,step=0.1,initial=1,label="No delta"),
-      steady=manipulate::checkbox(label="")
+      steady=manipulate::checkbox(TRUE,label="Steady?"),
+      Z.cv=manipulate::slider(0,0.2,step=0.05,initial=0,label="Z CV"),
+      No.cv=manipulate::slider(0,0.5,step=0.05,initial=0,label="No CV"),
       rerand=manipulate::button("Rerandomize")
     )
   }
-}  
-
+}
 
 
 ## Internal function to create the plot
-iCC.plot <- function(max.age,v,Zycl,ZdeltaAge,NodeltaAge,recruit.age,Z.param,No.param) {
+iCC.plot <- function(max.age,v,Zycl,deltaAge,recruit.age,Z.param,No.param) {
   # create ages to model over
   Ages <- 1:max.age
   # Make catches for the current slider selections (include possible violations)
   C.ass <- iCC.makeCatch(Ages,v,Zycl,recruit.age,Z.param,No.param)
   # Make catches for the default settings
-  Z.param.def <- list(mean=0.40,cv=0,delta=1,deltaAge=ZdeltaAge,dstdy=TRUE)
-  No.param.def <- list(mean=1000,cv=0,delta=1,deltaAge=NodeltaAge,dstdy=TRUE)
+  Z.param.def <- list(mean=0.40,cv=0,delta=1,deltaAge=deltaAge,dstdy=TRUE)
+  No.param.def <- list(mean=700,cv=0,delta=1,deltaAge=deltaAge,dstdy=TRUE)
   C.def <- iCC.makeCatch(Ages,v,Zycl,recruit.age,Z.param.def,No.param.def)
   # Make catches for the modified graph but wihout assumption violations
   # if no assumption violations, make same as actual so no extra line due to randomization
   if ((Z.param$delta==1) & (No.param$delta==1)) C.mod <- C.ass  
   else {
     Z.param.mod <- list(mean=Z.param$mean,cv=Z.param$cv,delta=1,
-                        deltaAge=ZdeltaAge,dstdy=TRUE)
+                        deltaAge=deltaAge,dstdy=TRUE)
     No.param.mod <- list(mean=No.param$mean,cv=No.param$cv,delta=1,
-                         deltaAge=NodeltaAge,dstdy=TRUE)    
+                         deltaAge=deltaAge,dstdy=TRUE)    
     C.mod <- iCC.makeCatch(Ages,v,Zycl,recruit.age,Z.param.mod,No.param.mod)
   }
   # Make the plot
-  op <- graphics::par(mar=c(3.5,3.5,1.5,1.5),mgp=c(2,0.4,0),tcl=-0.2,cex.main=0.75)
-  graphics::plot(Ages,log(C.def),type="o",col="gray",pch=19,lwd=3,
-                 ylab="log(Catch)",ylim=c(-3,8),main=iCC.title(Z.param,No.param))
+  op <- graphics::par(mar=c(3,3,1.5,1),mgp=c(1.9,0.4,0),tcl=-0.2,cex.main=1)
+  graphics::plot(Ages,log(C.def),type="o",col="gray",pch=16,lwd=3,
+                 ylab="log(Catch)",ylim=c(-3,6.2),main=iCC.title(Z.param,No.param))
   graphics::lines(Ages,log(C.ass),type="o",col="red",pch=19,lwd=3)
   graphics::lines(Ages,log(C.mod),type="o",col="blue",pch=19,lwd=3)
-  graphics::legend("topright",c("default","assumptions not met","if assumptions met"),
-                   lwd=2,pch=19,col=c("gray","red","blue"),cex=0.8)
+  graphics::legend("topright",c("assumptions not met","if assumptions met","default"),
+                   lwd=2,pch=16,col=c("red","blue","gray"),
+                   bg="white",box.col="white",inset=0.02)
   graphics::par(op)
 } # end iCC.plot