+++ NEWS.md
@@ -2,6 +2,7 @@
 * Compiling under R 3.2.3 and using roxygen2 5.0.1
 * DESCRIPTION: Updated.
 * Added Imports for `manipulate`, `relax`, `graphics`, and `stats`.  Added Suggests for `tcltk` (ultimately want to get rid of this and `relax`).
+* `catchCurveSim()`: Modified.  Major overhaul including changing to use `manipulate` rather than `relax`, removed `Zsteady=` and `NoSteady=` in favor of a checkbox (but must be the same for both Z and No), removed `ZDeltaAge=` and `NoDeltaAge=` in favor of one `deltaAge=` to be used for both Z and No.
 * `growthModelSim()`: Modified.  Fixed links to functions in `FSA` that had changed names.
 * `leslieSim()`: Modified.  Major overhaul including changing to use `manipulate` rather than `relax`, moved `removals` to an argument rather than a slider bar, chagned `type=` to `sim=`, and dramatically changed the plot for `sim="montecarlo"`.
 * `mrClosed1Sim()`: Modified.  Changed to use `manipulate` rather than `relax`.  Removed the `type=` argument and put it as a dynamic choice.  Added the `sim=` argument which allows simulations of the `distribution` or for assessing `assumptions` violations.  Changed look of the histogram (more like `hist()` from `FSA`).  Force more bins in the histogram.  Modified the legend look.  Changed default number of `rsmpls=`.  Streamlined code some.